📍 Feat: Add admin icon group api
